Move SSE4.1 and AVX2 drawing declarations to Drawing.Sprite.h
