+++ src/openrct2/drawing/Drawing.Sprite.cpp
@@ -20,6 +20,7 @@
 #include "../core/MemoryStream.h"
 #include "../core/Path.hpp"
 #include "../interface/ZoomLevel.h"
+#include "../platform/Platform.h"
 #include "../rct1/Csg.h"
 #include "../ui/UiContext.h"
 #include "Drawing.h"
@@ -310,6 +311,34 @@ static void OverrideElementOffsets(size_t index, G1Element& element)
             element.xOffset -= 1; // Steeplechase leftEighthToDiag angle 2
             break;
     }
+}
+
+static auto GetMaskFunction()
+{
+    if (Platform::AVX2Available())
+    {
+        LOG_VERBOSE("registering AVX2 mask function");
+        return MaskAvx2;
+    }
+    else if (Platform::SSE41Available())
+    {
+        LOG_VERBOSE("registering SSE4.1 mask function");
+        return MaskSse4_1;
+    }
+    else
+    {
+        LOG_VERBOSE("registering scalar mask function");
+        return MaskScalar;
+    }
+}
+
+static const auto MaskFunc = GetMaskFunction();
+
+void MaskFn(
+    int32_t width, int32_t height, const uint8_t* RESTRICT maskSrc, const uint8_t* RESTRICT colourSrc,
+    PaletteIndex* RESTRICT dst, int32_t maskWrap, int32_t colourWrap, int32_t dstWrap)
+{
+    MaskFunc(width, height, maskSrc, colourSrc, dst, maskWrap, colourWrap, dstWrap);
 }
 
 static void ReadAndConvertGxDat(IStream* stream, size_t count, bool is_rctc, G1Element* elements)
