+++ src/openrct2/drawing/SSE41Drawing.cpp
@@ -8,13 +8,12 @@
  *****************************************************************************/
 
 #include "../core/Guard.hpp"
-#include "Drawing.h"
+#include "Drawing.Sprite.h"
+#include "PaletteIndex.h"
 
 using OpenRCT2::Drawing::PaletteIndex;
 
 #ifdef __SSE4_1__
-
-    #include "Drawing.Sprite.h"
 
     #include <immintrin.h>
